refactor : User 디렉토리 정리
